requirements.txt
openai
langchain
huggingface_hub
langchain-openai
python-dotenv
streamlit
langchain-community
